feat: introduce normalizeCanvasData utility for canvas data handling
- Added normalizeCanvasData function to standardize canvas data by removing width and height properties and resetting collaborators.
- Updated App and BackupsDialog components to utilize the new utility for consistent canvas data normalization during scene updates.
